Remove custom getMaterialUIContext in favour of render override for tests
