+++ packages/react-ui/__tests__/widgets/TableWidgetUI.test.js
@@ -1,12 +1,10 @@
 import React from 'react';
 import { render, fireEvent, screen } from '../widgets/utils/testUtils';
 import TableWidgetUI from '../../src/widgets/TableWidgetUI/TableWidgetUI';
-import { getMaterialUIContext } from './testUtils';
 import { columns, rows } from '../../src/widgets/TableWidgetUI/mockData';
 
 describe('TableWidgetUI', () => {
-  const Widget = (props) =>
-    getMaterialUIContext(<TableWidgetUI columns={columns} rows={rows} {...props} />);
+  const Widget = (props) => <TableWidgetUI columns={columns} rows={rows} {...props} />;
 
   test('renders with default props', () => {
     render(<Widget />);
